deployment-service/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
httpx==0.28.1
pydantic==2.12.4
pydantic-settings==2.1.0
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
python-dotenv==1.2.1
aiokafka==0.12.0
